+++ sdk/communication/AzureCommunicationChat/Tests/ChatClientDVRTests.swift
@@ -24,7 +24,7 @@
 //
 // --------------------------------------------------------------------------
 
-import AzureCommunicationChat
+@testable import AzureCommunicationChat
 import AzureCommunicationCommon
 import AzureCore
 import AzureTest
@@ -48,7 +48,9 @@ class ChatClientDVRTests: RecordableXCTestCase<TestSettings> {
         let token = settings.token
         let credential = try CommunicationTokenCredential(token: token)
         let options = AzureCommunicationChatClientOptions(transportOptions: transportOptions)
+
         chatClient = try ChatClient(endpoint: endpoint, credential: credential, withOptions: options)
+        chatClient.registrationId = "0E0F0BE0-0000-00C0-B000-A00A00E00BD0"
     }
 
     func test_CreateThread_WithoutParticipants() {
@@ -105,6 +107,57 @@ class ChatClientDVRTests: RecordableXCTestCase<TestSettings> {
         waitForExpectations(timeout: 10.0) { error in
             if let error = error {
                 XCTFail("List threads timed out: \(error)")
+            }
+        }
+    }
+
+    func test_StartPushNotifications_ReturnsSuccess() {
+        let expectation = self.expectation(description: "Start push notifications")
+
+        chatClient
+            .startPushNotifications(deviceToken: "mockDeviceToken") { result in
+                switch result {
+                case let .success(response):
+                    XCTAssertEqual(response?.statusCode, 202)
+                    expectation.fulfill()
+                case .failure:
+                    XCTFail("Start push notifications failed.")
+                }
+            }
+
+        waitForExpectations(timeout: 10.0) { error in
+            if let error = error {
+                XCTFail("Start push notifications timed out: \(error)")
+            }
+        }
+    }
+
+    func test_StopPushNotifications_ReturnsSuccess() {
+        let expectation = self.expectation(description: "Stop push notifications")
+
+        // Start notifications first
+        chatClient
+            .startPushNotifications(deviceToken: "mockDeviceToken") { result in
+                switch result {
+                case .success:
+                    // Stop notifications
+                    self.chatClient.stopPushNotifications { result in
+                        switch result {
+                        case let .success(response):
+                            XCTAssertEqual(response?.statusCode, 202)
+                            expectation.fulfill()
+                        case .failure:
+                            XCTFail("Stop push notifications failed.")
+                        }
+                    }
+                case .failure:
+                    XCTFail("Start push notifications failed.")
+                }
+            }
+
+        waitForExpectations(timeout: 10.0) { error in
+            if let error = error {
+                XCTFail("Stop push notifications timed out: \(error)")
             }
         }
     }