+++ sdk/communication/AzureCommunicationChat/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.3.0-beta.1
+### New Features
+- `ChatClient` now supports Push Notifications for Chat events
+- Following methods added to `ChatClient`:
+  - `startPushNotifications(deviceToken:)`
+  - `stopPushNotifications()`
+- Added the prototol`PushNotificationKeyHandler` and the class `AppGroupPushNotificationKeyHandler` to support PushNotification Encryption Key Management
+
 ## 1.2.0 (2022-05-16)
 ### Features Added
 - Added two new events `realTimeNotificationConnected` and `realTimeNotificationDisconnected` that allow the developer to know when the connection to the real time notification server is active.